Move Merkle root computation to a helper module
Introduce a new helper module "merkle" which will contain utilities
for dealing with Merkle trees. Maybe some day we will use this as
common code with ProofListIndex.

The functions are made public so add proper documentation comments
for them.
